test(ingestion-consumer): parameterize the e2e matrix with rstest
Each harness test declares pin_stash and key_table cases and threads
the scheduler into the harness, replacing the two include-based test
binaries. The suite compiles once, cases filter by name, and stage 11
deletes one case line per test.
